placeholders are always less than results
