+++ autofit/non_linear/result.py
@@ -5,6 +5,26 @@ from autofit.mapper.prior_model.abstract import AbstractPriorModel
 from autofit.non_linear.samples import PDFSamples
 
 
+class Placeholder:
+    def __getattr__(self, item):
+        """
+        Placeholders return None to represent the missing result's value
+        """
+        return None
+
+    def __getstate__(self):
+        return {}
+
+    def __setstate__(self, state):
+        pass
+
+    def __gt__(self, other):
+        return False
+
+    def __lt__(self, other):
+        return True
+
+
 class Result:
     """
     @DynamicAttrs
@@ -31,6 +51,16 @@ class Result:
         self._instance = (
             samples.max_log_likelihood_instance if samples is not None else None
         )
+
+    def __gt__(self, other):
+        if isinstance(other, Placeholder):
+            return True
+        return self.log_likelihood > other.log_likelihood
+
+    def __lt__(self, other):
+        if isinstance(other, Placeholder):
+            return False
+        return self.log_likelihood < other.log_likelihood
 
     @property
     def log_likelihood(self):
